moved cy/measure over, about to try deleting pulse0
