+++ qiskit/providers/aer/pulse/de_solvers/pulse0_solvers.py
@@ -25,7 +25,7 @@ from scipy.linalg.blas import get_blas_funcs
 from .qiskit_zvode import qiskit_zvode
 
 # These pulse0 imports are only used by the Monte-Carlo solver
-from ..pulse0.cy.measure import occ_probabilities, write_shots_memory
+from ..qutip_extra_lite.cy.measure import occ_probabilities, write_shots_memory
 from ..qutip_extra_lite.cy.spmatfuncs import cy_expect_psi_csr, spmv_csr
 
 dznrm2 = get_blas_funcs("znrm2", dtype=np.float64)
